+++ Cql/CodeGeneration.NET/Toolkit/ElmToolkitResultRecord.cs
