+++ Cql/CodeGeneration.NET/Toolkit/Extensions/CqlToolkitExtensions.cs
