+++ Cql/CodeGeneration.NET/Toolkit/Internal/ElmToolkitScopedState.cs
