+++ Cql/Cql.CqlToElm/GlobalUsings.cs
@@ -11,13 +11,16 @@ global using System.Linq;
 global using System.Reflection;
 global using System.Text;
 global using System.Text.RegularExpressions;
-global using System.Threading.Tasks;
 global using System.Xml;
 global using System.Xml.Serialization;
-global using Microsoft.CodeAnalysis.CSharp.Syntax;
 global using Microsoft.Extensions.DependencyInjection;
 global using Microsoft.Extensions.Logging;
 global using Microsoft.Extensions.Options;
 global using NotNullAttribute = Antlr4.Runtime.Misc.NotNullAttribute;
 global using NotNullWhenAttribute = System.Diagnostics.CodeAnalysis.NotNullWhenAttribute;
 global using TypeInfo = Hl7.Cql.Model.TypeInfo;
+
+global using ElmLibrary = Hl7.Cql.Elm.Library;
+
+global using CqlToolkitConversionDictionary = System.Collections.Immutable.ImmutableDictionary<Hl7.Cql.Runtime.CqlVersionedLibraryIdentifier, Hl7.Cql.CqlToElm.Toolkit.CqlToolkitConversionRecord>;
+global using CqlToolkitConversionReadOnlyDictionary = System.Collections.Generic.IReadOnlyDictionary<Hl7.Cql.Runtime.CqlVersionedLibraryIdentifier, Hl7.Cql.CqlToElm.Toolkit.CqlToolkitConversionRecord>;
