+++ Cql/Cql.CqlToElm/Toolkit/CqlModel.cs
@@ -8,7 +8,7 @@
 
 using Hl7.Cql.Model;
 
-namespace CqlSdkPrototype.Cql;
+namespace Hl7.Cql.CqlToElm.Toolkit;
 
 /// <summary>
 /// Represents the model of CQL to use.
