+++ Cql/Cql.CqlToElm/Toolkit/CqlToolkitConfig.cs
@@ -6,11 +6,11 @@
  * available at https://raw.githubusercontent.com/FirelyTeam/firely-cql-sdk/main/LICENSE
  */
 
-using Hl7.Cql.CqlToElm;
+using System.Collections.Immutable;
 using Hl7.Cql.Model;
 using Hl7.Cql.Runtime;
 
-namespace CqlSdkPrototype.Cql;
+namespace Hl7.Cql.CqlToElm.Toolkit;
 
 /// <summary>
 /// The settings used to configure the CQL to ELM processor.
@@ -26,6 +26,9 @@ public sealed record CqlToolkitConfig(
     bool EnableIntervalDemotion = false,
     bool AllowNullIntervals = false)
 {
+    /// <summary>
+    /// The default configuration for the CQL toolkit.
+    /// </summary>
     public static CqlToolkitConfig Default { get; } = new();
 
     /// <summary>
