+++ Cql/Cql.CqlToElm/Toolkit/Internal/CqlToolkitServices.cs
@@ -6,14 +6,12 @@
  * available at https://raw.githubusercontent.com/FirelyTeam/firely-cql-sdk/main/LICENSE
  */
 
-using Hl7.Cql.Abstractions.Infrastructure;
-using Hl7.Cql.CqlToElm;
 using Hl7.Cql.CqlToElm.Visitors;
 using Hl7.Cql.Model;
 using Hl7.Cql.Runtime.Logging;
 using ExpressionVisitor = Hl7.Cql.CqlToElm.Visitors.ExpressionVisitor;
 
-namespace CqlSdkPrototype.Cql.Internal;
+namespace Hl7.Cql.CqlToElm.Toolkit.Internal;
 
 /// <summary>
 /// Services for translating CQL to ELM used by the <seealso cref="CqlToolkit"/>.
