+++ Cql/Cql.CqlToElm/Toolkit/Internal/LibraryBuilderProvider.cs
@@ -6,10 +6,9 @@
  * available at https://raw.githubusercontent.com/FirelyTeam/firely-cql-sdk/main/LICENSE
  */
 
-using Hl7.Cql.CqlToElm;
 using Hl7.Cql.Runtime;
 
-namespace CqlSdkPrototype.Cql.Internal;
+namespace Hl7.Cql.CqlToElm.Toolkit.Internal;
 
 /// <summary>
 /// Provides the implementation for <seealso cref="ILibraryProvider"/>
