+++ Cql/Cql.Packaging/GlobalUsings.cs
@@ -17,3 +17,7 @@ global using FhirModelCode = Hl7.Fhir.Model.Code;
 global using FhirResource = Hl7.Fhir.Model.Resource;
 global using Iso8601DateTimePrecision = Hl7.Cql.Iso8601.DateTimePrecision;
 global using SysDateTime = System.DateTime;
+
+global using ElmToolkitConversionDictionary = System.Collections.Immutable.ImmutableDictionary<Hl7.Cql.Runtime.CqlVersionedLibraryIdentifier, CqlSdkPrototype.Elm.ElmToolkitConversionRecord>;
+global using ElmToolkitConversionReadOnlyDictionary = System.Collections.Generic.IReadOnlyDictionary<Hl7.Cql.Runtime.CqlVersionedLibraryIdentifier, CqlSdkPrototype.Elm.ElmToolkitConversionRecord>;
+
