+++ Cql/CqlToElmTests/(tests)/TagsTest.cs
@@ -1,6 +1,6 @@
-using CqlSdkPrototype.Cql.Extensions;
 using CqlSdkPrototype.Elm.Extensions;
 using CqlSdkPrototype.Invocation.Extensions;
+using Hl7.Cql.CqlToElm.Toolkit.Extensions;
 
 namespace Hl7.Cql.CqlToElm.Test;
